+++ MobileClient/src/net/patchingzone/ru4real/sensors/AccelerometerListener.java
@@ -3,6 +3,6 @@ package net.patchingzone.ru4real.sensors;
 public interface AccelerometerListener {
 
 	public void onAccelerometerChanged(float x, float y, float z); 
-	public void onShake(float force); 
+	public void onShake(double force); 
 	
 }
